[quant] EqualizationQConfig to distinguish input/output activations
Summary: Created an EqualizationQConfig specifically for equalization.
This inherits from QConfig and is used to distinguish between inserting
an input observer with an output observer. Since the output observer
field is included in the EqualizationQConfig, we no longer need an
output observer field in the _InputEqualizationObserver

Test Plan: compiles

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 9c198e0797158cb6e5ab37e502610dbe232781f4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59739