[caffe2] add an EstimateAllBlobSizes operator
This operator is similar to `GetAllBlobNames` but also returns the estimated
size required to serialize each node.

One goal of this operator is to allow checkpoint saving logic to estimate the
amount of space/bandwidth required to save a checkpoint when first starting
training, without actually serializing any blobs yet.  Currently the
checkpointing logic uses `GetAllBlobNames` to determine the blobs to
checkpoint.  It can instead be updated to use `EstimateAllBlobSizes` to also
get an estimate for how much space will be required for the checkpoint.

Differential Revision: [D29020227](https://our.internmc.facebook.com/intern/diff/D29020227/)

ghstack-source-id: 131045068
Pull Request resolved: https://github.com/pytorch/pytorch/pull/59775